CON-737: Fix exception mechanism when filesystem sizes exceeed limits
# Changes
It appears that with FatFs we currently have some limits of sizes due to the usage of FAT32.
When the user creates an enclave with size too small, the enclave just crashes with the a C++ error.

Also, we found out that when changing the size of the filesystem, we would get basically the same issue,  i.e. a C++ exception not caught.
It would be better to get a proper Java/Kotlin exception and this is what we are trying to do here.

With this PR we would get this for the size limit
```
        ... 20 more
Caused by: java.lang.RuntimeException: 
Unable to initialize the enclave's persistent filesystem.
Have you inadvertently changed the sizes of the persistent filesystem in the enclave's config
 or the content of the filesystem file?
Have you changed enclave mode?
        at com.oracle.svm.jni.JNIJavaCallWrappers.jniInvoke_ARRAY:Ljava_lang_RuntimeException_2_0002e_0003cinit_0003e_00028Ljava_lang_String_2_00029V(JNIJavaCallWrappers.java:0)
        at com.oracle.svm.jni.functions.JNIFunctions$NewObjectWithObjectArrayArgFunctionPointer.invoke(JNIFunctions.java)
        at com.oracle.svm.jni.functions.JNIFunctions.ThrowNew(JNIFunctions.java:818)
        at com.r3.conclave.enclave.internal.Native.setupFileSystems(Native.java)
        at com.r3.conclave.enclave.internal.NativeEnclaveEnvironment.setupFileSystems(NativeEnclaveEnvironment.kt:115)
        at com.r3.conclave.enclave.Enclave.setupFileSystems(Enclave.kt:275)
        at com.r3.conclave.enclave.Enclave.access$setupFileSystems(Enclave.kt:56)
        at com.r3.conclave.enclave.Enclave$AdminHandler.onOpen(Enclave.kt:411)
        at com.r3.conclave.enclave.Enclave$AdminHandler.onReceive(Enclave.kt:368)

```
and this for the size changes


```
Caused by: java.lang.RuntimeException: Unable to initialize the enclave's persistent filesystem
        at com.oracle.svm.jni.JNIJavaCallWrappers.jniInvoke_ARRAY:Ljava_lang_RuntimeException_2_0002e_0003cinit_0003e_00028Ljava_lang_String_2_00029V(JNIJavaCallWrappers.java:0)
        at com.oracle.svm.jni.functions.JNIFunctions$NewObjectWithObjectArrayArgFunctionPointer.invoke(JNIFunctions.java)
        at com.oracle.svm.jni.functions.JNIFunctions.ThrowNew(JNIFunctions.java:818)
        at com.r3.conclave.enclave.internal.Native.setupFileSystems(Native.java)
        at com.r3.conclave.enclave.internal.NativeEnclaveEnvironment.setupFileSystems(NativeEnclaveEnvironment.kt:115)
        at com.r3.conclave.enclave.Enclave.setupFileSystems(Enclave.kt:275)
        at com.r3.conclave.enclave.Enclave.access$setupFileSystems(Enclave.kt:56)
        at com.r3.conclave.enclave.Enclave$AdminHandler.onOpen(Enclave.kt:411)
...
```


# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?
- [ ] If the gradle plugin has been updated, have the gradle plugin tests been run in TeamCity?

+++ conclave-host/src/main/kotlin/com/r3/conclave/host/internal/fatfs/FileSystemHandler.kt
@@ -55,13 +55,13 @@ class FileSystemHandler(enclaveFileSystemFilePaths: List<Path>) {
 
     @Suppress("unused")
     @Synchronized
-    fun read(drive: Int, sectorId: Int, numSectors: Int, sectorSize: Int): ByteArray {
+    fun read(drive: Int, sectorId: Long, numSectors: Int, sectorSize: Int): ByteArray {
         val randomAccessFile = files[drive]
 
         val position = sectorId * sectorSize
         val readSize = numSectors * sectorSize
         val buffer = ByteArray(readSize)
-        randomAccessFile.seek(position.toLong())
+        randomAccessFile.seek(position)
         randomAccessFile.readFully(buffer)
         return buffer
     }
@@ -69,7 +69,7 @@ class FileSystemHandler(enclaveFileSystemFilePaths: List<Path>) {
 
     @Suppress("unused")
     @Synchronized
-    fun write(drive: Int, inputBuffer: ByteArray, sectorSize: Int, indices: IntArray): Int {
+    fun write(drive: Int, inputBuffer: ByteArray, sectorSize: Int, indices: LongArray): Int {
         val randomAccessFile = files[drive]
 
         var bufferIt = 0
